Fix compilation of views containing relative imports
