+++ web/src/test/scala/quasar/api/services/Fixture.scala
@@ -19,18 +19,16 @@ package quasar.api.services
 import slamdata.Predef._
 import quasar.api.{ResponseOr, ResponseT}
 import quasar.contrib.pathy._
-import quasar.effect._
 import quasar.fp._
 import quasar.fp.free._
-import quasar.fs._
 import quasar.fs.InMemory.InMemState
 import quasar.fs.mount._
-import quasar.fs.mount.cache.VCache
 import quasar.metastore.{MetaStore, MetaStoreFixture}
 import quasar.api.JsonFormat.{SingleArray, LineDelimited}
 import quasar.api.JsonPrecision.{Precise, Readable}
 import quasar.api.MessageFormat.JsonContentType
 import quasar.main._
+import quasar.main.Fixture._
 import quasar.server.qErrsToResponseT
 
 import argonaut.{Json, Argonaut}
@@ -87,68 +85,6 @@ object Fixture {
         EntityEncoder.stringEncoder(Charset.`UTF-8`).toEntity(csv.value)
       }
   }
-
-  def mountingInter(mounts: Map[APath, MountConfig]): Task[Mounting ~> Task] =
-    mountingInterInspect(mounts).map{ case (inter, ref) => inter }
-
-  def mountingInterInspect(mounts: Map[APath, MountConfig]): Task[(Mounting ~> Task, Task[Map[APath, MountConfig]])] = {
-    type MEff[A] = Coproduct[Task, MountConfigs, A]
-    TaskRef(mounts).map { configsRef =>
-
-      val mounter: Mounting ~> Free[MEff, ?] = Mounter.trivial[MEff]
-
-      val meff: MEff ~> Task =
-        reflNT[Task] :+: KeyValueStore.impl.fromTaskRef(configsRef)
-
-      (foldMapNT(meff) compose mounter, configsRef.read)
-    }
-  }
-
-  def inMemFS(
-    state: InMemState = InMemState.empty,
-    mounts: MountingsConfig = MountingsConfig.empty
-  ): Task[FS] =
-    inMemFSInspect(state, mounts).map { case (inter, ref) => inter }
-
-  def inMemFSInspect(
-    state: InMemState = InMemState.empty,
-    mounts: MountingsConfig = MountingsConfig.empty
-  ): Task[(FS, Task[(InMemState, Map[APath, MountConfig])])] = {
-    val noShutdown: Task[Unit] = Task.now(())
-    (InMemory.runBackendInspect(state) |@| mountingInterInspect(mounts.toMap))((fsAndRef, mountAndRef) =>
-      (FS(
-        fsAndRef._1 andThen injectFT[Task, QErrs_Task],
-        mountAndRef._1 andThen injectFT[Task, QErrs_Task],
-        noShutdown), (fsAndRef._2 |@| mountAndRef._2).tupled))
-  }
-
-  def inMemFSEvalInspect(
-    state: InMemState = InMemState.empty,
-    mounts: MountingsConfig = MountingsConfig.empty,
-    metaRefT: Task[TaskRef[MetaStore]] = MetaStoreFixture.createNewTestMetastore().flatMap(TaskRef(_)),
-    persist: quasar.db.DbConnectionConfig => MainTask[Unit] = _ => ().point[MainTask]
-  ): Task[(CoreEffIO ~> QErrs_TaskM, Task[(InMemState, Map[APath, MountConfig])])] =
-    for {
-      r         <- TaskRef(Tags.Min(Option.empty[VCache.Expiration]))
-      metaRef   <- metaRefT
-      result    <- inMemFSInspect(state, mounts)
-      (fs, ref) = result
-      eval      <- CoreEff.defaultImpl(fs, metaRef, persist)
-    } yield
-      (injectFT[Task, QErrs_CRW_Task] :+: eval andThen
-      foldMapNT(
-        (Read.fromTaskRef(r) andThen injectFT[Task, QErrs_Task])  :+:
-        (Write.fromTaskRef(r) andThen injectFT[Task, QErrs_Task]) :+:
-        injectFT[Task, QErrs_Task]                                :+:
-        injectFT[QErrs, QErrs_Task]), ref)
-
-  def inMemFSEval(
-     state: InMemState = InMemState.empty,
-     mounts: MountingsConfig = MountingsConfig.empty,
-     metaRefT: Task[TaskRef[MetaStore]] = MetaStoreFixture.createNewTestMetastore().flatMap(TaskRef(_)),
-     persist: quasar.db.DbConnectionConfig => MainTask[Unit] = _ => ().point[MainTask]
-   ): Task[CoreEffIO ~> QErrs_TaskM] =
-    inMemFSEvalInspect(state, mounts, metaRefT, persist).map{ case (inter, ref) => inter }
 
   def inMemFSWeb(
     state: InMemState = InMemState.empty,
